currently working on this NN model
This notebook shows how to read numpy array file in trunks and trains the neural network for large dataset. The code is still runing but currently have 0.3 AP for 1 epcho, the model is currently runing using all split data (train on 230M trainining data , and test on 56M validation data), but I want to share this code so everyone could work on and take a look on how to train the model using all data in trunks.